requirements.txt
paramiko==2.7.1
pywin32==227
bypy==1.6.10
numpy==1.18.1
selenium==3.141.0
requests==2.18.4
opencv_python==4.4.0.44
Pillow~=7.1.2
win32gui==221.6

configparser~=5.0.0
opencv-python~=4.4.0.44
tqdm~=4.50.2
pandas~=0.25.3
matplotlib~=3.2.1
scipy~=1.5.2
pyinstaller~=4.0
xlwt~=1.3.0
pclpy~=0.11.0
open3d~=0.11.2
pyntcloud~=0.1.3